requirements.txt
langchain
python-dotenv
streamlit
openai
langchain_openai
langchain_core
langchain_community
langchain_groq
faiss-cpu
pypdf